Remove unnecessary decimal rescaling - memory usage optimization (#160)
* Remove unnecessary decimal rescaling

* Move benchmarks to separate file, add new benchmark tests
